Move imgui-filebrowser implementation into separate file
This is to avoid polluting the global namespace with macros from Windows.h.

Also explicitly include <imgui.h> instead of checking that it has been
included before.
